🚚 Separate search index generation into its own file
